move sound classes to registry package
