+++ src/main/java/com/github/thedeathlycow/frostiful/registry/FBlockSoundGroups.java
@@ -1,4 +1,4 @@
-package com.github.thedeathlycow.frostiful.sound;
+package com.github.thedeathlycow.frostiful.registry;
 
 import net.minecraft.sound.BlockSoundGroup;
