+++ src/main/java/com/github/thedeathlycow/frostiful/registry/FBlocks.java
@@ -2,7 +2,6 @@ package com.github.thedeathlycow.frostiful.registry;
 
 import com.github.thedeathlycow.frostiful.Frostiful;
 import com.github.thedeathlycow.frostiful.block.*;
-import com.github.thedeathlycow.frostiful.sound.FBlockSoundGroups;
 import net.minecraft.block.*;
 import net.minecraft.block.enums.NoteBlockInstrument;
 import net.minecraft.block.piston.PistonBehavior;
